% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress_dir}
\alias{compress_dir}
\title{compress a dir}
\usage{
compress_dir(type = c("project", "system"), dir = NULL)
}
\description{
@description compress_dir()  creates .tar.gz file
   if type is system, zipped file will be placed in toGoogle dir
   if type is project, zipped file will be placed in compressed folder in
   this package.
 @param type project, if dir is inside project; 
             system, if the_dirs is list of complete diretory paths
 @param dir directory 
 @return 
 @export
}
