% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_files.R
\name{create_fake_files}
\alias{create_fake_files}
\title{@title create_fake_files
 
 @description  create fake dir,  numbered fake file names.}
\usage{
create_fake_files()
}
\description{
@return character vector of file names.
 @export
}
