% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{create_S3_object}
\alias{create_S3_object}
\title{create S3 object}
\usage{
create_S3_object(s = NULL, class = NULL)
}
\value{
S3 object of class
}
\description{
Create two simple classes, \code{secret} and \code{top_secret}
}
\details{
create_S3_object
}
