% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_fake_files.R
\name{create_sandbox}
\alias{create_sandbox}
\title{reate_sandbox}
\usage{
create_sandbox()
}
\value{
character vector of full path fake file names.
}
\description{
create a sandbox; fake dir,  numbered fake file names.
}
