% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_files.R
\name{.is_sandbox}
\alias{.is_sandbox}
\title{.is_sandbox
@description
Logical.  True if \code{the_dir} is sandbox.  Otherwise false.
@param:
the_dir.  String. Directory to check
@export}
\usage{
.is_sandbox(the_dir = NULL)
}
\description{
.is_sandbox
@description
Logical.  True if \code{the_dir} is sandbox.  Otherwise false.
@param:
the_dir.  String. Directory to check
@export
}
