% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_files.R
\name{.is_valid_path}
\alias{.is_valid_path}
\title{.is_valid_path}
\usage{
.is_valid_path(the_dir = NULL)
}
\arguments{
\item{the_dir}{Directory provided by user.}
}
\description{
Logical  Checks if path exists.  HELPER
}
\details{
Note \code{.} prefix.  This or \verb{keyword  internal} means not exported,
internal only AND will not be included in documentation.
}
