% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_yaml_header.R
\name{fetch_yaml_header}
\alias{fetch_yaml_header}
\title{@title fetch_yaml_header
 @description  This function does most of actual work:  Given a file name,
 return yaml header as character vector, one element per line.
 @param f name of file
 @return character vector, one entry for each yaml line, `---` inclusive.
 @export}
\usage{
fetch_yaml_header(f)
}
\description{
@title fetch_yaml_header
 @description  This function does most of actual work:  Given a file name,
 return yaml header as character vector, one element per line.
 @param f name of file
 @return character vector, one entry for each yaml line, `---` inclusive.
 @export
}
