% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{foo}
\alias{foo}
\title{execuable chunk}
\usage{
foo()
}
\description{
execuable chunk
}
\details{
Details

\if{html}{\out{<div class="sourceCode r">}}\preformatted{1+1
#> [1] 2
}\if{html}{\out{</div>}}
}
