% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{foo2}
\alias{foo2}
\title{getNamespace("")}
\usage{
foo2()
}
\description{
Add R chunk

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ls(getNamespace("jimTools"))
#>  [1] "compress_dir"                    "create_S3_object"               
#>  [3] "create_sandbox"                  "dir"                            
#>  [5] "dt"                              "e1"                             
#>  [7] "e2"                              "f"                              
#>  [9] "f_object"                        "file"                           
#> [11] "foo"                             "foo2"                           
#> [13] "foo3"                            "get_compressed_file_name"       
#> [15] "get_info"                        "get_options"                    
#> [17] "get_prefix"                      "info"                           
#> [19] "jim_str"                         "jr_get_env"                     
#> [21] "jr_install.packages"             "jr_package_dependencies"        
#> [23] "jr_package_dependencies_reverse" "jr_set_env"                     
#> [25] "jr_update_packages"              "log_sandbox"                    
#> [27] "old"                             "populate_sandbox"               
#> [29] "remove_sandbox"                  "report_info"                    
#> [31] "reveal"                          "reveal.secret"                  
#> [33] "reveal.top_secret"               "server"                         
#> [35] "set_env_name"                    "show_env"                       
#> [37] "ui"                              "what_is_it"                     
#> [39] "where"                           "z"
}\if{html}{\out{</div>}}
}
