% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{foo3}
\alias{foo3}
\title{foo3}
\usage{
foo3()
}
\description{
No errors, but no sign of it when document

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cat("#' @export\\n")
#> #' @export
}\if{html}{\out{</div>}}
}
