% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_file_names.R
\name{get_files}
\alias{get_files}
\title{@title get_files
  @param path Path to directory.
  @param pattern Use pattern (regex) to return subset of files.  If pattern is NULL, all files
 in path will be returned.}
\usage{
get_files(path = NULL, pattern = NULL)
}
\description{
@title get_files
  @param path Path to directory.
  @param pattern Use pattern (regex) to return subset of files.  If pattern is NULL, all files
 in path will be returned.
}
