% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{get_prefix}
\alias{get_prefix}
\title{get_prefix}
\usage{
get_prefix()
}
\value{

}
\description{
prepares date-based prefix for final compressed file.
}
\keyword{backup}
\keyword{compress}
\keyword{helper}
