% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_file_names.R
\name{get_proposed_name}
\alias{get_proposed_name}
\title{get_proposed_name}
\usage{
get_proposed_name(the_files = NULL, prefix = NULL)
}
\arguments{
\item{the_files}{File names that will receive new prefix.}

\item{prefix}{Proposed new prefix.}
}
\description{
attach proposed prefix.
}
