% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_RMD_files}
\alias{get_RMD_files}
\title{@title get_RMD_files}
\usage{
get_RMD_files(path = ".", pattern = NULL, recursive = FALSE)
}
\description{
@description  Given a path and pattern, return char vector of file names.
 Ignores files that begin with '_'
 @param path  Directory to query. 
 @param pattern regex pattern to match files 
 @return 
 @export
}
