% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tags_line.R
\name{get_TAGS_line}
\alias{get_TAGS_line}
\title{@title get_tags_line
 @description Given a yaml header (everything between `---` inclusive),
 return the line begins with `TAGS:`  Return NA if no line contains
 `TAGS:`
 @param yaml char vector with each element one line of yaml header
 @keywords yaml
 @export}
\usage{
get_TAGS_line(header = null)
}
\description{
@title get_tags_line
 @description Given a yaml header (everything between `---` inclusive),
 return the line begins with `TAGS:`  Return NA if no line contains
 `TAGS:`
 @param yaml char vector with each element one line of yaml header
 @keywords yaml
 @export
}
