% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040_jim_str.R
\name{jim_str}
\alias{jim_str}
\title{jim_str() show str() information}
\usage{
jim_str(X = NULL, max.level = 1, vec.len = 1)
}
\arguments{
\item{X}{object}

\item{max.level}{number of levels}

\item{vec.len}{number of  elements to show at each level}
}
\value{

}
\description{
friendly version of str()
}
\keyword{internal}
