% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimTools-package.R
\docType{package}
\name{jimTools-package}
\alias{jimTools}
\alias{jimTools-package}
\title{jimTools: Misc Tools I Find Useful}
\description{
Misc tools I find useful.
}
\details{
This file (jimTools-package.R) should signal documentation to included functions.
}
\author{
\strong{Maintainer}: jim Last \email{jim.rothstein@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
