% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_help_file.R
\docType{package}
\name{jimTools}
\alias{jimTools}
\title{jimTools: A package for odds & ends I have found useful.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

\section{To Use}{

?jimTools  at the command line.
}

