% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_packages.R
\name{jr_install.packages}
\alias{jr_install.packages}
\title{file <- "tools_packages.R' jr_install.packages"}
\usage{
jr_install.packages(pkg_names = NULL)
}
\arguments{
\item{pkg_names}{packages to install}
}
\description{
file <- "tools_packages.R' jr_install.packages"
}
