% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_packages.R
\name{jr_package_dependencies}
\alias{jr_package_dependencies}
\title{jr_package_dependencies}
\usage{
jr_package_dependencies(package = NULL)
}
\arguments{
\item{package}{package or packages needs what?}
}
\description{
jr_package_dependencies
}
