% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_packages.R
\name{jr_package_dependencies_reverse}
\alias{jr_package_dependencies_reverse}
\title{jr_package_dependencies_reverse}
\usage{
jr_package_dependencies_reverse(package = NULL)
}
\arguments{
\item{package}{Which package or packages depend on this one?}
}
\description{
jr_package_dependencies_reverse
}
