% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_packages.R
\name{jr_set_env}
\alias{jr_set_env}
\title{jr_set_env}
\usage{
jr_set_env(key = NULL, value = NULL)
}
\arguments{
\item{key}{key is name of env variable}

\item{value}{value of env variable}
}
\description{
HELPER
}
