% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_IO.R
\name{read_file_names}
\alias{read_file_names}
\title{read_file_names}
\usage{
read_file_names(path = NULL, pattern = NULL)
}
\arguments{
\item{path}{Path to directory.}

\item{pattern}{Use pattern (regex) to return subset of files.  If pattern is NULL, all files
in path will be returned.}
}
\value{
character vector of file names
}
\description{
Returns character vector of files in path that match regex
 pattern.
}
