% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{remove_prefix}
\alias{remove_prefix}
\title{remove_prefix
 @param s  is single character vector or string char[1], where s is of the
  from "X: a,b,d..."  and X:  (TAGS:) is to be removed.
 @return  char[1]  without X:}
\usage{
remove_prefix(s = character())
}
\description{
remove_prefix
 @param s  is single character vector or string char[1], where s is of the
  from "X: a,b,d..."  and X:  (TAGS:) is to be removed.
 @return  char[1]  without X:
}
