% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_files.R
\name{remove_sandbox}
\alias{remove_sandbox}
\title{remove_sandbox(the_dir)}
\usage{
remove_sandbox(the_dir)
}
\arguments{
\item{path}{Path (directory) of files to be changed.}
}
\description{
Removes sandbox, all tmp directories and all files.
}
