% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_printing.R
\name{ren_file}
\alias{ren_file}
\title{ren_html (file,dir)
render rmd to html}
\usage{
ren_file(
  file,
  output_format = "html_document",
  output_dir = "~/Downloads/print_and_delete"
)
}
\arguments{
\item{file}{file must be sent in the form here(dir, file)}

\item{output_format}{default "html_document" also "pdf_document"}

\item{outdir_dir}{default: "~/Downloads/print_and_delete"}
}
\value{

}
\description{
ren_html (file,dir)
render rmd to html
}
