% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_printing.R
\name{ren_github}
\alias{ren_github}
\title{ren_github(file, dir)}
\usage{
ren_github(x = NULL, output_dir = "./md")
}
\arguments{
\item{file}{file name}

\item{dir}{directory, use as here(file, dir)}
}
\value{

}
\description{
render rmd to .md github format
}
\details{
====================================


 No details.
}
\examples{
 ren_github(x=here("rmd", "001_compress.Rmd"))
 ren_github(x=here("rmd", "001_compress.Rmd"), 
   output_dir = "./md")
 ren_github(x=here("rmd", "001_compress.Rmd"), 
   output_dir = "~/Downloads/print_and_delete")
 sum(1:10)
 \dontrun{
   sum("a")
   }
}
\seealso{
many places for more info.
}
