% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{ren_pdf}
\alias{ren_pdf}
\title{ren_github(file,dir)
render rmd to pdf}
\usage{
ren_pdf(file, dir)
}
\arguments{
\item{file}{file name}

\item{dir}{directory, use as here(file, dir)}
}
\value{

}
\description{
ren_github(file,dir)
render rmd to pdf
}
