% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_IO.R
\name{rename_file_names}
\alias{rename_file_names}
\title{rename_file_names}
\usage{
rename_file_names(path = NULL, from = NULL, to = NULL)
}
\arguments{
\item{from}{Original file names (still on disk).}

\item{to}{Proposed new file names.}

\item{Full}{path to directory where files are located.}
}
\description{
Renames all files that have changed file name.
}
