% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_file_names.R
\name{rename_files}
\alias{rename_files}
\title{rename_files}
\usage{
rename_files(path = NULL, from = NULL, to = NULL)
}
\arguments{
\item{from}{Original file names (still on disk).}

\item{to}{Proposed new file names.}

\item{Full}{path to directory where files are located.}
}
\description{
rename_files
}
