% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{reveal}
\alias{reveal}
\title{generic function \code{reveal}}
\usage{
reveal(x)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{create generic, "reveal" 
}\if{html}{\out{</div>}}
}
