% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_file_names.R
\name{save_old_names}
\alias{save_old_names}
\title{@title save_old_names
 @description
 safely save copy of original files in separate environmentt}
\usage{
save_old_names(the_files = NULL)
}
\description{
@title save_old_names
 @description
 safely save copy of original files in separate environmentt
}
