% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_files.R
\name{write_fake_file}
\alias{write_fake_file}
\title{write_fake_file}
\usage{
write_fake_file(f)
}
\arguments{
\item{f}{the fake file that wil be written to disk.}
}
\description{
Adds minimal content stub to fake file and writes to disk.
}
\details{
if directory does not exist = errror;
}
