% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{compress_dir}
\alias{compress_dir}
\title{compress a dir}
\usage{
compress_dir(
  prefix = NULL,
  out_file = NULL,
  type = c("project", "system"),
  dir = NULL
)
}
\arguments{
\item{prefix}{of file name}

\item{out_file}{prefix plus this will be name of  compressed file.}

\item{type}{project, if dir is inside R project; (ie relative dir), OR
system, if the_dirs is list of complete diretory paths}

\item{dir}{directory}
}
\value{

}
\description{
compress_dir()  creates .tar.gz file
if type is system, zipped file will be placed in toGoogle dir
if type is project, zipped file will be placed in compressed folder in
this package.
}
\keyword{backup}
\keyword{compress}
