% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_example.R
\name{foo2}
\alias{foo2}
\title{getNamespace("")}
\usage{
foo2()
}
\description{
Add R chunk

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ls(getNamespace("jimTools"))
#>  [1] "compress_dir"                    "create_S3_object"               
#>  [3] "create_sandbox"                  "dir"                            
#>  [5] "e1"                              "e2"                             
#>  [7] "f"                               "f_object"                       
#>  [9] "file"                            "foo"                            
#> [11] "foo2"                            "foo3"                           
#> [13] "get_compressed_file_name"        "get_info"                       
#> [15] "get_options"                     "get_prefix"                     
#> [17] "info"                            "jim_str"                        
#> [19] "jr_get_env"                      "jr_install.packages"            
#> [21] "jr_package_dependencies"         "jr_package_dependencies_reverse"
#> [23] "jr_set_env"                      "jr_update_packages"             
#> [25] "log_sandbox"                     "populate_sandbox"               
#> [27] "remove_sandbox"                  "report_info"                    
#> [29] "reveal"                          "reveal.secret"                  
#> [31] "reveal.top_secret"               "set_env_name"                   
#> [33] "show_env"                        "what_is_it"                     
#> [35] "where"
}\if{html}{\out{</div>}}
}
