% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{get_compressed_file_name}
\alias{get_compressed_file_name}
\title{get_compressed_file_name}
\usage{
get_compressed_file_name(prefix = NULL, dir = NULL)
}
\arguments{
\item{prefix}{prefix name}

\item{dir}{directory}
}
\value{

}
\description{
Given a prefix and dir, returns name of compressed file.
}
\keyword{backup}
\keyword{compress}
\keyword{helper}
