% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_get_RMD_files.R
\name{get_RMD_files}
\alias{get_RMD_files}
\title{@description  Given a path and pattern, return char vector of file names.
 Ignores files that begin with '_'
 @param path  Directory to query. 
 @param pattern regex pattern to match files 
 @return full file name (with directory)}
\usage{
get_RMD_files(path = ".", pattern = NULL, recursive = FALSE)
}
\description{
@description  Given a path and pattern, return char vector of file names.
 Ignores files that begin with '_'
 @param path  Directory to query. 
 @param pattern regex pattern to match files 
 @return full file name (with directory)
}
\keyword{@export}
\keyword{internal}
