% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_help_file.R
\name{jimTools}
\alias{jimTools}
\title{jimTools: A package for odds & ends I have found useful.}
\description{
jimTools: Package for functions and info I find useful.
}
\section{Foo functions}{

The foo functions ...
}

\section{To Use}{

?jimTools  at the command line.
ls(asNamespace("jimTools")) - returns chr[] of functions
lsf.str("package:jimTools") - list functions
}

