% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing.R
\name{ren_github}
\alias{ren_github}
\title{====================================
ren_github(file, dir)
render rmd to .md github format}
\usage{
ren_github(file, dir)
}
\arguments{
\item{file}{file name}

\item{dir}{directory, use as here(file, dir)}
}
\value{

}
\description{
====================================
ren_github(file, dir)
render rmd to .md github format
}
