% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_files.R
\name{write_fake_file}
\alias{write_fake_file}
\title{adds content to a sandbox file}
\usage{
write_fake_file(file)
}
\arguments{
\item{file}{the fake file that wil be written to disk.}
}
\description{
Adds minimal content stub to sandbox file and writes to disk.
}
\details{
if directory does not exist = errror;
}
